Move common 'manifest' pkg -> state/manifest
